Revert "release 1.2.0"
