+++ PLDroidPlayerDemo/.gitignore
@@ -1,8 +1,6 @@
-*.iml
 .gradle
 /local.properties
 /.idea/workspace.xml
 /.idea/libraries
 .DS_Store
 /build
-/captures
