+++ PLDroidPlayerDemo/app/build.gradle
@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "24.0.0 rc1"
+    compileSdkVersion 22
+    buildToolsVersion "22.0.1"
 
     defaultConfig {
-        applicationId "com.pili.pldroid.playerdemo"
+        applicationId "com.pili.android.playerdemo"
         minSdkVersion 9
-        targetSdkVersion 23
+        targetSdkVersion 22
         versionCode 1
         versionName "1.0"
     }
@@ -20,6 +20,8 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:design:23.2.1'
-    compile files('libs/pldroid-player-1.2.0.jar')
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.android.support:appcompat-v7:22.0.0'
+    compile files('libs/ijkmediaplayer.jar')
+    compile files('libs/pldroid-player-1.1.6.jar')
 }
