--- PLDroidPlayerDemo/app/src/main/java/com/pili/pldroid/playerdemo/PLAudioPlayerActivity.java
@@ -1,193 +1,0 @@
-package com.pili.pldroid.playerdemo;
-
-import android.content.Context;
-import android.media.AudioManager;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.View;
-
-import com.pili.pldroid.player.PLMediaPlayer;
-
-import java.io.IOException;
-
-/**
- *  This demo shows how to use PLMediaPlayer API playing audio stream
- */
-public class PLAudioPlayerActivity extends AppCompatActivity {
-
-    private static final String TAG = PLAudioPlayerActivity.class.getSimpleName();
-
-    private PLMediaPlayer mMediaPlayer;
-    private String mAudioPath;
-    private View mLoadingView;
-    private boolean mIsStopped = false;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_audio_player);
-        mLoadingView = findViewById(R.id.LoadingView);
-        mAudioPath = getIntent().getStringExtra("videoPath");
-
-        AudioManager audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        audioManager.requestAudioFocus(null, AudioManager.STREAM_MUSIC, AudioManager.AUDIOFOCUS_GAIN);
-        mMediaPlayer = new PLMediaPlayer();
-        mMediaPlayer.setOnPreparedListener(mOnPreparedListener);
-        mMediaPlayer.setOnCompletionListener(mOnCompletionListener);
-        mMediaPlayer.setOnErrorListener(mOnErrorListener);
-        mMediaPlayer.setOnInfoListener(mOnInfoListener);
-        mMediaPlayer.setOnBufferingUpdateListener(mOnBufferingUpdateListener);
-
-        prepare();
-    }
-
-    @Override
-    protected void onDestroy() {
-        release();
-        AudioManager audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        audioManager.abandonAudioFocus(null);
-        super.onDestroy();
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        mMediaPlayer.start();
-    }
-
-    @Override
-    protected void onPause() {
-        mMediaPlayer.pause();
-        super.onPause();
-    }
-
-    public void onClickPlay(View v) {
-        if (mIsStopped) {
-            prepare();
-        } else {
-            mMediaPlayer.start();
-        }
-    }
-
-    public void onClickPause(View v) {
-        mMediaPlayer.pause();
-    }
-
-    public void onClickResume(View v) {
-        mMediaPlayer.start();
-    }
-
-    public void onClickStop(View v) {
-        mMediaPlayer.stop();
-        mMediaPlayer.reset();
-        mIsStopped = true;
-    }
-
-    public void release() {
-        mMediaPlayer.stop();
-        mMediaPlayer.release();
-        mMediaPlayer = null;
-    }
-
-    private void prepare() {
-        try {
-            mMediaPlayer.setDataSource(mAudioPath);
-            mMediaPlayer.prepareAsync();
-        } catch (IllegalArgumentException e) {
-            e.printStackTrace();
-        } catch (IllegalStateException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private PLMediaPlayer.OnPreparedListener mOnPreparedListener = new PLMediaPlayer.OnPreparedListener() {
-        @Override
-        public void onPrepared(PLMediaPlayer mp) {
-            Log.i(TAG, "On Prepared !");
-            mMediaPlayer.start();
-            mIsStopped = false;
-        }
-    };
-
-    private PLMediaPlayer.OnInfoListener mOnInfoListener = new PLMediaPlayer.OnInfoListener() {
-        @Override
-        public boolean onInfo(PLMediaPlayer mp, int what, int extra) {
-            Log.i(TAG, "OnInfo, what = " + what + ", extra = " + extra);
-            switch (what) {
-                case PLMediaPlayer.MEDIA_INFO_BUFFERING_START:
-                    mLoadingView.setVisibility(View.VISIBLE);
-                    break;
-                case PLMediaPlayer.MEDIA_INFO_BUFFERING_END:
-                case PLMediaPlayer.MEDIA_INFO_AUDIO_RENDERING_START:
-                    mLoadingView.setVisibility(View.GONE);
-                    break;
-                default:
-                    break;
-            }
-            return true;
-        }
-    };
-
-    private PLMediaPlayer.OnBufferingUpdateListener mOnBufferingUpdateListener = new PLMediaPlayer.OnBufferingUpdateListener() {
-        @Override
-        public void onBufferingUpdate(PLMediaPlayer mp, int percent) {
-            Log.d(TAG, "onBufferingUpdate: " + percent + "%");
-        }
-    };
-
-    /**
-     *  Listen the event of playing complete
-     *  For playing local file, it's called when reading the file EOF
-     *  For playing network stream, it's called when the buffered bytes played over
-     *
-     *  If setLooping(true) is called, the player will restart automatically
-     *  And ｀onCompletion｀ will not be called
-     */
-    private PLMediaPlayer.OnCompletionListener mOnCompletionListener = new PLMediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(PLMediaPlayer mp) {
-            Log.d(TAG, "Play Completed !");
-        }
-    };
-
-    private PLMediaPlayer.OnErrorListener mOnErrorListener = new PLMediaPlayer.OnErrorListener() {
-        @Override
-        public boolean onError(PLMediaPlayer mp, int errorCode) {
-            Log.e(TAG, "Error happened, errorCode = " + errorCode);
-            switch (errorCode) {
-                case PLMediaPlayer.ERROR_CODE_INVALID_URI:
-                    break;
-                case PLMediaPlayer.ERROR_CODE_404_NOT_FOUND:
-                    break;
-                case PLMediaPlayer.ERROR_CODE_CONNECTION_REFUSED:
-                    break;
-                case PLMediaPlayer.ERROR_CODE_CONNECTION_TIMEOUT:
-                    break;
-                case PLMediaPlayer.ERROR_CODE_EMPTY_PLAYLIST:
-                    break;
-                case PLMediaPlayer.ERROR_CODE_STREAM_DISCONNECTED:
-                    break;
-                case PLMediaPlayer.MEDIA_ERROR_UNKNOWN:
-                    break;
-                default:
-                    break;
-            }
-            // Todo pls handle the error status here, retry or call finish()
-            finish();
-            // The PLMediaPlayer has moved to the Error state, if you want to retry, must reset first !
-            // try {
-            //     mMediaPlayer.reset();
-            //     mMediaPlayer.setDataSource(mAudioPath);
-            //     mMediaPlayer.prepareAsync();
-            // } catch (IOException e) {
-            //     e.printStackTrace();
-            // }
-            // Return true means the error has been handled
-            // If return false, then `onCompletion` will be called
-            return true;
-        }
-    };
-}
