--- PLDroidPlayerDemo/app/src/main/java/com/pili/pldroid/playerdemo/PLMediaPlayerActivity.java
@@ -1,290 +1,0 @@
-package com.pili.pldroid.playerdemo;
-
-import android.content.Context;
-import android.media.AudioManager;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.Gravity;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
-import android.view.View;
-import android.view.WindowManager;
-import android.widget.FrameLayout;
-import android.widget.Toast;
-
-import com.pili.pldroid.player.AVOptions;
-import com.pili.pldroid.player.PLMediaPlayer;
-
-import java.io.IOException;
-
-/**
- *  This demo shows how to use PLMediaPlayer API playing video stream
- */
-public class PLMediaPlayerActivity extends AppCompatActivity {
-
-    private static final String TAG = PLMediaPlayerActivity.class.getSimpleName();
-
-    private SurfaceView mSurfaceView;
-    private PLMediaPlayer mMediaPlayer;
-    private View mLoadingView;
-    private AVOptions mAVOptions;
-
-    private int mSurfaceWidth  = 0;
-    private int mSurfaceHeight = 0;
-
-    private String mVideoPath = null;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_media_player);
-        mLoadingView = findViewById(R.id.LoadingView);
-        mSurfaceView = (SurfaceView) findViewById(R.id.SurfaceView);
-        mSurfaceView.getHolder().addCallback(mCallback);
-
-        mVideoPath = getIntent().getStringExtra("videoPath");
-
-        mAVOptions = new AVOptions();
-
-        if (isLiveStreaming(mVideoPath)) {
-            // the unit of timeout is ms
-            mAVOptions.setInteger(AVOptions.KEY_GET_AV_FRAME_TIMEOUT, 10 * 1000);
-            // Some optimization with buffering mechanism when be set to 1
-            mAVOptions.setInteger(AVOptions.KEY_LIVE_STREAMING, 1);
-        }
-
-        // 1 -> hw codec enable, 0 -> disable [recommended]
-        int codec = getIntent().getIntExtra("mediaCodec", 0);
-        mAVOptions.setInteger(AVOptions.KEY_MEDIACODEC, codec);
-
-        AudioManager audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        audioManager.requestAudioFocus(null, AudioManager.STREAM_MUSIC, AudioManager.AUDIOFOCUS_GAIN);
-    }
-
-    @Override
-    protected void onDestroy() {
-        release();
-        AudioManager audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        audioManager.abandonAudioFocus(null);
-        super.onDestroy();
-    }
-
-    public void onClickPlay(View v) {
-        prepare();
-    }
-
-    public void onClickPause(View v) {
-        if (mMediaPlayer != null) {
-            mMediaPlayer.pause();
-        }
-    }
-
-    public void onClickResume(View v) {
-        if (mMediaPlayer != null) {
-            mMediaPlayer.start();
-        }
-    }
-
-    public void onClickStop(View v) {
-        if (mMediaPlayer != null) {
-            mMediaPlayer.stop();
-            mMediaPlayer.reset();
-        }
-    }
-
-    public void release() {
-        if (mMediaPlayer != null) {
-            mMediaPlayer.stop();
-            mMediaPlayer.release();
-            mMediaPlayer = null;
-        }
-    }
-
-    private void prepare() {
-
-        if (mMediaPlayer != null) {
-            release();
-        }
-
-        try {
-            mMediaPlayer = new PLMediaPlayer(mAVOptions);
-
-            mMediaPlayer.setOnPreparedListener(mOnPreparedListener);
-            mMediaPlayer.setOnVideoSizeChangedListener(mOnVideoSizeChangedListener);
-            mMediaPlayer.setOnCompletionListener(mOnCompletionListener);
-            mMediaPlayer.setOnErrorListener(mOnErrorListener);
-            mMediaPlayer.setOnInfoListener(mOnInfoListener);
-            mMediaPlayer.setOnBufferingUpdateListener(mOnBufferingUpdateListener);
-
-            // set replay if completed
-            // mMediaPlayer.setLooping(true);
-
-            mMediaPlayer.setDataSource(mVideoPath);
-            mMediaPlayer.setDisplay(mSurfaceView.getHolder());
-            mMediaPlayer.prepareAsync();
-
-        } catch (IllegalArgumentException e) {
-            e.printStackTrace();
-        } catch (IllegalStateException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private SurfaceHolder.Callback mCallback = new SurfaceHolder.Callback() {
-
-        @Override
-        public void surfaceCreated(SurfaceHolder holder) {
-            prepare();
-        }
-
-        @Override
-        public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
-            mSurfaceWidth = width;
-            mSurfaceHeight = height;
-        }
-
-        @Override
-        public void surfaceDestroyed(SurfaceHolder holder) {
-            release();
-        }
-    };
-
-    private PLMediaPlayer.OnVideoSizeChangedListener mOnVideoSizeChangedListener = new PLMediaPlayer.OnVideoSizeChangedListener() {
-        public void onVideoSizeChanged(PLMediaPlayer mp, int width, int height) {
-            Log.i(TAG, "onVideoSizeChanged, width = "+ width + ",height = " + height);
-            // resize the display window to fit the screen
-            if (width != 0 && height != 0) {
-                float ratioW = (float) width/(float) mSurfaceWidth;
-                float ratioH = (float) height/(float) mSurfaceHeight;
-                float ratio = Math.max(ratioW, ratioH);
-                width  = (int) Math.ceil((float)width/ratio);
-                height = (int) Math.ceil((float)height/ratio);
-                FrameLayout.LayoutParams layout = new FrameLayout.LayoutParams(width, height);
-                layout.gravity = Gravity.CENTER;
-                mSurfaceView.setLayoutParams(layout);
-            }
-        }
-    };
-
-    private PLMediaPlayer.OnPreparedListener mOnPreparedListener = new PLMediaPlayer.OnPreparedListener() {
-        @Override
-        public void onPrepared(PLMediaPlayer mp) {
-            Log.i(TAG, "On Prepared !");
-            mMediaPlayer.start();
-        }
-    };
-
-    private PLMediaPlayer.OnInfoListener mOnInfoListener = new PLMediaPlayer.OnInfoListener() {
-        @Override
-        public boolean onInfo(PLMediaPlayer mp, int what, int extra) {
-            Log.i(TAG, "OnInfo, what = " + what + ", extra = " + extra);
-            switch (what) {
-                case PLMediaPlayer.MEDIA_INFO_BUFFERING_START:
-                    mLoadingView.setVisibility(View.VISIBLE);
-                    break;
-                case PLMediaPlayer.MEDIA_INFO_BUFFERING_END:
-                case PLMediaPlayer.MEDIA_INFO_VIDEO_RENDERING_START:
-                    mLoadingView.setVisibility(View.GONE);
-                    break;
-                default:
-                    break;
-            }
-            return true;
-        }
-    };
-
-    private PLMediaPlayer.OnBufferingUpdateListener mOnBufferingUpdateListener = new PLMediaPlayer.OnBufferingUpdateListener() {
-        @Override
-        public void onBufferingUpdate(PLMediaPlayer mp, int percent) {
-            Log.d(TAG, "onBufferingUpdate: " + percent + "%");
-        }
-    };
-
-    /**
-     *  Listen the event of playing complete
-     *  For playing local file, it's called when reading the file EOF
-     *  For playing network stream, it's called when the buffered bytes played over
-     *
-     *  If setLooping(true) is called, the player will restart automatically
-     *  And ｀onCompletion｀ will not be called
-     *
-     */
-    private PLMediaPlayer.OnCompletionListener mOnCompletionListener = new PLMediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(PLMediaPlayer mp) {
-            Log.d(TAG, "Play Completed !");
-            showToastTips("Play Completed !");
-            finish();
-        }
-    };
-
-    private PLMediaPlayer.OnErrorListener mOnErrorListener = new PLMediaPlayer.OnErrorListener() {
-        @Override
-        public boolean onError(PLMediaPlayer mp, int errorCode) {
-            Log.e(TAG, "Error happened, errorCode = " + errorCode);
-            switch (errorCode) {
-                case PLMediaPlayer.ERROR_CODE_INVALID_URI:
-                    showToastTips("Invalid URL !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_404_NOT_FOUND:
-                    showToastTips("404 resource not found !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_CONNECTION_REFUSED:
-                    showToastTips("Connection refused !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_CONNECTION_TIMEOUT:
-                    showToastTips("Connection timeout !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_EMPTY_PLAYLIST:
-                    showToastTips("Empty playlist !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_STREAM_DISCONNECTED:
-                    showToastTips("Stream disconnected !");
-                    break;
-                case PLMediaPlayer.ERROR_CODE_IO_ERROR:
-                    showToastTips("Network IO Error !");
-                    break;
-                case PLMediaPlayer.MEDIA_ERROR_UNKNOWN:
-                default:
-                    showToastTips("unknown error !");
-                    break;
-            }
-            // Todo pls handle the error status here, retry or call finish()
-            finish();
-            // The PLMediaPlayer has moved to the Error state, if you want to retry, must reset first !
-            // try {
-            //     mMediaPlayer.reset();
-            //     mMediaPlayer.setDisplay(mSurfaceView.getHolder());
-            //     mMediaPlayer.setDataSource(mVideoPath);
-            //     mMediaPlayer.prepareAsync();
-            // } catch (IOException e) {
-            //     e.printStackTrace();
-            // }
-            // Return true means the error has been handled
-            // If return false, then `onCompletion` will be called
-            return true;
-        }
-    };
-
-    private void showToastTips(final String tips) {
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                Toast.makeText(PLMediaPlayerActivity.this, tips, Toast.LENGTH_LONG).show();
-            }
-        });
-    }
-
-    private boolean isLiveStreaming(String url) {
-        if (url.startsWith("rtmp://")
-                || (url.startsWith("http://") && url.endsWith(".m3u8"))
-                || (url.startsWith("http://") && url.endsWith(".flv"))) {
-            return true;
-        }
-        return false;
-    }
-}
