+++ PLDroidPlayerDemo/app/src/main/java/com/pili/pldroid/playerdemo/widget/AspectLayout.java
@@ -15,7 +15,6 @@ import com.pili.pldroid.player.common.Util;
  * Created by jerikc on 15/11/22.
  */
 public class AspectLayout extends RelativeLayout {
-
     private static final String TAG = "AspectLayout";
 
     private int mWidthMeasureSpec;
