+++ PLDroidPlayerDemo/app/src/main/java/com/pili/pldroid/playerdemo/widget/MediaController.java
@@ -27,13 +27,11 @@ import com.pili.pldroid.player.IMediaController;
 import java.util.Locale;
 
 /**
- * You can write a custom MediaController instead of this class
- * A MediaController widget must implement all the interface defined by com.pili.pldroid.player.IMediaController
+ * Created by jerikc on 15/5/29.
  */
 public class MediaController extends FrameLayout implements IMediaController {
-
-    private static final String TAG = "PLMediaController";
-    private IMediaController.MediaPlayerControl mPlayer;
+    private static final String TAG = "MyMediaController";
+    private MediaPlayerControl mPlayer;
     private Context mContext;
     private PopupWindow mWindow;
     private int mAnimStyle;
@@ -75,6 +73,7 @@ public class MediaController extends FrameLayout implements IMediaController {
     private Runnable mLastSeekBarRunnable;
     private boolean mDisableProgress = false;
 
+
     public MediaController(Context context, AttributeSet attrs) {
         super(context, attrs);
         mRoot = this;
@@ -110,7 +109,6 @@ public class MediaController extends FrameLayout implements IMediaController {
     public void onFinishInflate() {
         if (mRoot != null)
             initControllerView(mRoot);
-        super.onFinishInflate();
     }
 
     private void initFloatingWindow() {
@@ -377,7 +375,7 @@ public class MediaController extends FrameLayout implements IMediaController {
             if (!fromuser)
                 return;
 
-            final int newposition = (int) (mDuration * progress) / 1000;
+            final long newposition = (mDuration * progress) / 1000;
             String time = generateTime(newposition);
             if (mInstantSeeking) {
                 mHandler.removeCallbacks(mLastSeekBarRunnable);
@@ -395,7 +393,7 @@ public class MediaController extends FrameLayout implements IMediaController {
 
         public void onStopTrackingTouch(SeekBar bar) {
             if (!mInstantSeeking)
-                mPlayer.seekTo((int)(mDuration * bar.getProgress()) / 1000);
+                mPlayer.seekTo((mDuration * bar.getProgress()) / 1000);
 
             show(sDefaultTimeout);
             mHandler.removeMessages(SHOW_PROGRESS);
@@ -405,9 +403,9 @@ public class MediaController extends FrameLayout implements IMediaController {
         }
     };
 
-    private OnClickListener mRewListener = new OnClickListener() {
+    private View.OnClickListener mRewListener = new View.OnClickListener() {
         public void onClick(View v) {
-            int pos = (int)mPlayer.getCurrentPosition();
+            long pos = mPlayer.getCurrentPosition();
             pos -= 5000; // milliseconds
             mPlayer.seekTo(pos);
             setProgress();
@@ -416,9 +414,9 @@ public class MediaController extends FrameLayout implements IMediaController {
         }
     };
 
-    private OnClickListener mFfwdListener = new OnClickListener() {
+    private View.OnClickListener mFfwdListener = new View.OnClickListener() {
         public void onClick(View v) {
-            int pos = (int)mPlayer.getCurrentPosition();
+            long pos = mPlayer.getCurrentPosition();
             pos += 15000; // milliseconds
             mPlayer.seekTo(pos);
             setProgress();
@@ -530,7 +528,7 @@ public class MediaController extends FrameLayout implements IMediaController {
         if (mShowing) {
             if (mAnchor != null) {
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.ICE_CREAM_SANDWICH) {
-                    //mAnchor.setSystemUiVisibility(View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
+                    mAnchor.setSystemUiVisibility(View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
                 }
             }
             try {
