+++ PLDroidPlayerDemo/app/src/main/res/layout/activity_audio_player.xml
@@ -1,40 +1,53 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:background="@color/background_material_dark"
+    tools:context="com.pili.pldroid.playerdemo.AudioPlayerActivity">
 
-    <ImageView
+    <!--<com.pili.pldroid.player.widget.VideoView-->
+        <!--android:id="@+id/video_view"-->
+        <!--android:layout_width="fill_parent"-->
+        <!--android:layout_height="fill_parent"-->
+        <!--android:layout_gravity="center" />-->
+
+    <Button
+        android:id="@+id/back_btn"
+        android:text="@string/back_title"
+        android:layout_gravity="top|left"
+        android:background="@color/background_floating_material_dark"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/audio"
-        android:layout_margin="20dp"/>
+        android:layout_height="wrap_content" />
 
     <LinearLayout
-        android:id="@+id/LoadingView"
-        android:layout_width="match_parent"
+        android:id="@+id/buffering_indicator"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:gravity="center"
-        android:orientation="horizontal">
-        <ProgressBar
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="5dp"/>
+        android:focusable="false"
+        android:focusableInTouchMode="false"
+        android:orientation="vertical"
+        android:visibility="gone" >
+
         <TextView
+            android:id="@+id/buffering_msg"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            android:text="buffering"
-            android:textColor="@android:color/black"
-            android:textSize="20sp"/>
+            android:layout_gravity="center"
+            android:adjustViewBounds="true"
+            android:ellipsize="marquee"
+            android:marqueeRepeatLimit="0"
+            android:scaleType="centerInside"
+            android:singleLine="true"
+            android:text="@string/buffering"
+            android:textSize="12sp" />
+
+        <ProgressBar
+            android:id="@+id/buffering_progress"
+            style="@android:style/Widget.ProgressBar.Large"
+            android:layout_width="60dip"
+            android:layout_height="60dip" />
     </LinearLayout>
-
-    <include
-        layout="@layout/layout_buttons"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="10dp"
-        android:layout_gravity="bottom" />
-
-</FrameLayout>
+</RelativeLayout>
