--- PLDroidPlayerDemo/app/src/main/res/layout/activity_pl_video_texture.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.pili.pldroid.player.widget.PLVideoTextureView
-        android:id="@+id/VideoView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"/>
-
-    <LinearLayout
-        android:id="@+id/LoadingView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:orientation="horizontal">
-        <ProgressBar
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="5dp"/>
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            android:text="@string/buffering"
-            android:textColor="@android:color/white"
-            android:textSize="20sp"/>
-    </LinearLayout>
-    
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_gravity="top|right"
-        android:layout_margin="20dp">
-
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_rotate_right"
-            android:layout_margin="5dp"
-            android:onClick="onClickRotate"/>
-
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_switch_screen"
-            android:layout_margin="5dp"
-            android:onClick="onClickSwitchScreen"/>
-
-    </LinearLayout>
-
-</FrameLayout>
