--- PLDroidPlayerDemo/app/src/main/res/layout/activity_video_view.xml
@@ -1,65 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/background_material_light" >
-
-    <com.pili.pldroid.playerdemo.widget.AspectLayout
-        android:id="@+id/aspect_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <com.pili.pldroid.player.widget.VideoView
-            android:id="@+id/video_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_centerInParent="true"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentRight="true"
-            />
-    </com.pili.pldroid.playerdemo.widget.AspectLayout>
-
-    <Button
-        android:id="@+id/back_btn"
-        android:text="@string/back_title"
-        android:layout_gravity="top|left"
-        android:background="@color/background_floating_material_dark"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp" />
-
-    <EditText
-        android:layout_alignParentBottom="true"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-    <LinearLayout
-        android:id="@+id/buffering_indicator"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:focusable="false"
-        android:focusableInTouchMode="false"
-        android:orientation="vertical"
-        android:visibility="gone"
-        android:layout_centerInParent="true">
-        <TextView
-            android:id="@+id/buffering_msg"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:adjustViewBounds="true"
-            android:ellipsize="marquee"
-            android:marqueeRepeatLimit="0"
-            android:scaleType="centerInside"
-            android:singleLine="true"
-            android:text="@string/buffering"
-            android:textSize="12sp" />
-        <ProgressBar
-            android:id="@+id/buffering_progress"
-            style="@android:style/Widget.ProgressBar.Large"
-            android:layout_width="60dip"
-            android:layout_height="60dip" />
-    </LinearLayout>
-
-</RelativeLayout>
