--- PLDroidPlayerDemo/app/src/main/res/layout/dialog_setting.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/decode_config"
-        android:layout_margin="10dp"
-        android:textSize="15sp"
-        android:textStyle="bold" />
-
-    <Spinner
-        android:id="@+id/CodecSpinner"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="10dp"
-        android:layout_marginBottom="20dp" />
-
-</LinearLayout>
